[mod] paginas y componentes
