--- pages/api/auth/[...nextauth].js
@@ -1,12 +1,0 @@
-import NextAuth from 'next-auth'
-import EmailProvider from "next-auth/providers/email"
-
-export default NextAuth({
-    providers: [
-        // Passwordless / email sign in
-        EmailProvider({
-            server: process.env.EMAIL_SERVER,
-            from: process.env.EMAIL_FROM
-        }),
-    ]
-})
